+++ baselines/common/vec_env/dummy_vec_env.py
@@ -1,6 +1,6 @@
 import numpy as np
 from gym import spaces
-from . import VecEnv
+from .vec_env import VecEnv
 from .util import copy_obs_dict, dict_to_obs, obs_space_info
 
 class DummyVecEnv(VecEnv):
